+++ phpBB2/IBProArcade.php
@@ -30,6 +30,10 @@ include_once($phpbb_root_path . 'includes/constants_arcade.'.$phpEx);
 $userdata			= session_pagestart($user_ip, PAGE_ACTIVITY); 
 init_userprefs($userdata); 
 include_once($phpbb_root_path . 'includes/functions_arcade.'.$phpEx);
+if (!phpbb_request_source_is_same_origin())
+{
+	message_die(GENERAL_ERROR, 'Cross-site score submissions are not accepted.');
+}
 $arcade_version = $arcade->arcade_config('version');
 //
 // End session management 
@@ -46,10 +50,11 @@ foreach($HTTP_POST_VARS as $key => $value)
 {
   if(!is_array($value))
   {
-    $log .= $key . '=>' . $value . ' ';
+	$log .= substr(str_replace(array("\r", "\n", "\0"), '', (string) $key), 0, 64) . '=>'
+		. substr(str_replace(array("\r", "\n", "\0"), '', (string) $value), 0, 512) . ' ';
   }
 }
-$log = mysqli_real_escape_string($db->db_connect_id, substr($log, 0, 4000));
+$log = $db->sql_escape(substr($log, 0, 4000));
 $sql = "INSERT INTO " . iNA_LOG . " (user_id, name, value, date) 
   VALUES (".(int) $userdata['user_id'].", 'GAME', '$log', ".time().")";
 $db->sql_query($sql);
@@ -58,12 +63,13 @@ $db->sql_query($sql);
 //
 if($do == 'verifyscore')
 {
-  $randchar1 = rand(1, 200);
-	$randchar2 = rand(1, 200);
+	$randchar1 = (hexdec(bin2hex(phpbb_random_bytes(2))) % 200) + 1;
+	$randchar2 = (hexdec(bin2hex(phpbb_random_bytes(2))) % 200) + 1;
+	$session_hash_sql = $db->sql_escape($session_info['arcade_hash']);
 
   $sql = "UPDATE " . iNA_SESSIONS . "
     SET randchar1 = $randchar1, randchar2 = $randchar2
-    WHERE arcade_hash = '". $session_info['arcade_hash'] . "'";
+    WHERE arcade_hash = '". $session_hash_sql . "'";
   if(!$result = $db->sql_query($sql)) 
   {
   	message_die(GENERAL_ERROR, $lang['session_data_error'] . $lang['newscore_close'], "", __LINE__, __FILE__, $sql); 
@@ -79,7 +85,9 @@ else if ($do == 'savescore' || $do == 'newscore')
 
   $decodescore = $arcade->score * $session_info['randchar1'] ^ $session_info['randchar2'];
   
-  if(is_finite((float) $arcade->score) && $enscore == $decodescore)
+  if(is_finite((float) $arcade->score) && abs((float) $arcade->score) <= 9999999999.9999 &&
+	(int) $session_info['randchar1'] >= 1 && (int) $session_info['randchar1'] <= 200 &&
+	(int) $session_info['randchar2'] >= 1 && (int) $session_info['randchar2'] <= 200 && $enscore == $decodescore)
   {  
     $arcade->game_name = $session_info['game_name'];